Remove client-side custom dimension validation.
This behaviour doesn't match what is happening server-side, as evidenced
by having 20+ char custom dimension column names.

Rather than try and have the client library match the canonical
validation which is happening on the server, just allow the server to
validate the request and surface any errors in HTTP responses.